modularized graphql schema, renamed some more things, moved datasource configuration out of index.js, added rudimentary readme
