+++ graphqlSchema/typeDefs.js
@@ -1,0 +1,6 @@
+const rootSchema = require('./rootSchema')
+const placeholder = require('./placeholder')
+
+const schemaArray = [rootSchema, placeholder]
+
+module.exports = schemaArray
