--- resolvers.js
@@ -1,9 +1,0 @@
-const queryResolvers = require('./resolvers/query/queryResolvers')
-const mutationResolvers = require('./resolvers/mutation/mutationResolvers')
-
-module.exports = () => {
-  return {
-    ...queryResolvers,
-    ...mutationResolvers,
-  }
-}
