+++ resolvers/mutation/mutationResolvers.js
@@ -1,7 +1,7 @@
 module.exports = {
   Mutation: {
 		async placeholder(parent, { input }, { dataSources, req, app, postgres }){
-			return await dataSources.database.mutationPlaceholder('placeholder')
+			return await dataSources.placeholderDatabase.mutationPlaceholder('placeholder')
 		},
 
 		async placeholderApi(parent, { input }, { dataSources, req, app, postgres }){
