+++ resolvers/query/queryResolvers.js
@@ -1,11 +1,11 @@
 module.exports = {
   Query: {
 		async placeholder(parent, { input }, { dataSources, req, app, postgres }){
-			return await dataSources.database.queryPlaceholder('placeholder')
+			return await dataSources.placeholderDatabase.queryPlaceholder('placeholder')
 		},
 
 		async placeholderApi(parent, { input }, { dataSources, req, app, postgres }){
-			return await dataSources.placeholderApi.queryPlaceholder('placeholder')
+			return await dataSources.placeholderApi.queryPlaceholder('placeholder Api')
 		},
   },
 }
