+++ resolvers/resolvers.js
@@ -1,0 +1,9 @@
+const queryResolvers = require('./query/queryResolvers')
+const mutationResolvers = require('./mutation/mutationResolvers')
+
+module.exports = () => {
+  return {
+    ...queryResolvers,
+    ...mutationResolvers,
+  }
+}
